+++ src/arpack/eigs_sym_sp_d.cc
@@ -35,8 +35,9 @@ RTensor eigs(const RSparse &A, EigType eig_type, size_t neig,
        */
     return eigs_small(full(A), eig_type, neig, eigenvectors, converged);
   }
-  return eigs([&](const RTensor &x) { return mmult(A, x); }, A.columns(),
-              eig_type, neig, eigenvectors, converged);
+  return eigs([&](const RTensor &x) { return mmult(A, x); },
+              static_cast<size_t>(A.columns()), eig_type, neig, eigenvectors,
+              converged);
 }
 
 }  // namespace linalg
