+++ src/linalg/eig_power_sp_d.cc
@@ -32,8 +32,8 @@ namespace linalg {
 double eig_power_right(const RSparse &O, RTensor *vector, size_t iter,
                        double tol) {
   assert(O.rows() == O.columns());
-  return eig_power([&O](const RTensor &x) { return mmult(O, x); }, O.columns(),
-                   vector, iter, tol);
+  return eig_power([&O](const RTensor &x) { return mmult(O, x); },
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 /**Left eigenvalue and eigenvector with the largest absolute
@@ -48,7 +48,7 @@ double eig_power_left(const RSparse &O, RTensor *vector, size_t iter,
   assert(O.rows() == O.columns());
   auto OT = transpose(O);
   return eig_power([&OT](const RTensor &x) { return mmult(OT, x); },
-                   O.columns(), vector, iter, tol);
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 }  // namespace linalg
