+++ src/linalg/eig_power_sp_z.cc
@@ -32,8 +32,8 @@ namespace linalg {
 tensor::cdouble eig_power_right(const CSparse &O, CTensor *vector, size_t iter,
                                 double tol) {
   assert(O.rows() == O.columns());
-  return eig_power([&O](const CTensor &x) { return mmult(O, x); }, O.columns(),
-                   vector, iter, tol);
+  return eig_power([&O](const CTensor &x) { return mmult(O, x); },
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 /**Left eigenvalue and eigenvector with the largest absolute
@@ -48,7 +48,7 @@ tensor::cdouble eig_power_left(const CSparse &O, CTensor *vector, size_t iter,
   assert(O.rows() == O.columns());
   auto OT = transpose(O);
   return eig_power([&OT](const CTensor &x) { return mmult(OT, x); },
-                   O.columns(), vector, iter, tol);
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 }  // namespace linalg
