+++ src/tensor/tensor_sum_d.cc
@@ -22,7 +22,7 @@
 namespace tensor {
 
 double sum(const RTensor &r) {
-  return std::accumulate(r.begin_const(), r.end_const(), (double)0.0);
+  return std::accumulate(r.begin_const(), r.end_const(), 0.0);
 }
 
 }  // namespace tensor
